feat: implement scoped catalog foundation for story 2.1

+++ api/admin/internal/logic/pms/product_brand/deleteproductbrandlogic.go
@@ -2,21 +2,17 @@ package product_brand
 
 import (
 	"context"
+	"github.com/feihua/zero-admin/api/admin/internal/common"
 	"github.com/feihua/zero-admin/api/admin/internal/common/errorx"
+	"github.com/feihua/zero-admin/api/admin/internal/common/res"
 	"github.com/feihua/zero-admin/api/admin/internal/svc"
 	"github.com/feihua/zero-admin/api/admin/internal/types"
 	"github.com/feihua/zero-admin/rpc/pms/pmsclient"
 	"github.com/zeromicro/go-zero/core/logc"
-	"google.golang.org/grpc/status"
-
 	"github.com/zeromicro/go-zero/core/logx"
+	"google.golang.org/grpc/status"
 )
 
-// DeleteProductBrandLogic 删除商品品牌
-/*
-Author: LiuFeiHua
-Date: 2025/05/26 10:33:54
-*/
 type DeleteProductBrandLogic struct {
 	logx.Logger
 	ctx    context.Context
@@ -24,27 +20,22 @@ type DeleteProductBrandLogic struct {
 }
 
 func NewDeleteProductBrandLogic(ctx context.Context, svcCtx *svc.ServiceContext) *DeleteProductBrandLogic {
-	return &DeleteProductBrandLogic{
-		Logger: logx.WithContext(ctx),
-		ctx:    ctx,
-		svcCtx: svcCtx,
-	}
+	return &DeleteProductBrandLogic{Logger: logx.WithContext(ctx), ctx: ctx, svcCtx: svcCtx}
 }
-
-// DeleteProductBrand 删除商品品牌
-func (l *DeleteProductBrandLogic) DeleteProductBrand(req *types.DeleteProductBrandReq) (resp *types.BaseResp, err error) {
-	_, err = l.svcCtx.ProductBrandService.DeleteProductBrand(l.ctx, &pmsclient.DeleteProductBrandReq{
-		Ids: req.Ids,
-	})
-
+func (l *DeleteProductBrandLogic) DeleteProductBrand(req *types.DeleteProductBrandReq) (*types.BaseResp, error) {
+	userId, err := common.GetUserId(l.ctx)
+	if err != nil {
+		return nil, err
+	}
+	writeScope, err := common.ResolveWriteGovernanceScope(l.ctx, common.RequestedGovernanceScope{ScopeType: req.ScopeType, PlatformID: req.PlatformId, TenantID: req.TenantId, MerchantID: req.MerchantId})
+	if err != nil {
+		return nil, errorx.NewDefaultError(err.Error())
+	}
+	_, err = l.svcCtx.ProductBrandService.DeleteProductBrand(l.ctx, &pmsclient.DeleteProductBrandReq{Ids: req.Ids, UpdateBy: userId, Scope: common.PMSGovernanceScope(writeScope)})
 	if err != nil {
 		logc.Errorf(l.ctx, "删除商品品牌失败,参数：%+v,响应：%s", req, err.Error())
 		s, _ := status.FromError(err)
 		return nil, errorx.NewDefaultError(s.Message())
 	}
-
-	return &types.BaseResp{
-		Code:    "000000",
-		Message: "删除商品品牌成功",
-	}, nil
+	return res.Success()
 }
